requirements.txt
# ARPGuard Application Dependencies

# Core Dependencies
PyQt5>=5.15.9
scapy>=2.4.5
netifaces>=0.11.0
requests>=2.31.0
pyyaml>=6.0.1
psutil>=5.9.0

# UI Components
pyqtgraph>=0.13.3
pyqtwebengine>=5.15.6

# Packet Analysis
dpkt>=1.9.8
# pypcap>=1.3.0  # Removed as it requires WinPcap/Npcap
dnslib>=0.9.19

# Utility Libraries
python-dateutil>=2.8.2
pyfiglet>=1.0.0
tabulate>=0.9.0
colorama>=0.4.4
termcolor>=2.0.1

# Reporting
jinja2>=3.1.0
markdown>=3.5.0
weasyprint>=60.0
beautifulsoup4>=4.12.0

# Data Storage
sqlalchemy>=2.0.0
alembic>=1.12.0

# Security & Cryptography
cryptography>=41.0.0
bcrypt>=4.0.0

# Machine Learning dependencies
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0
tensorflow==2.13.0
keras==2.13.1
torch>=1.9.0
torchvision>=0.10.0
shap==0.41.0
xgboost==1.7.3
lightgbm==3.3.5
joblib>=1.3.0

# Testing Dependencies
pytest>=7.4.0
pytest-qt>=4.1.0
pytest-cov>=4.1.0
coverage>=7.3.0
mock>=4.0.3

# Development Tools
pre-commit>=2.17.0
pylint>=2.12.2
mypy>=0.941
black>=22.1.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Utilities
tqdm>=4.66.0
psutil==5.9.5

# Web/API
fastapi==0.100.0
uvicorn==0.23.1
python-dotenv==1.0.0
websockets>=10.4
python-jose>=3.3.0
pydantic>=2.0.0
fastapi-jwt-auth>=0.5.0
httpx>=0.24.0

# Core dependencies
mkdocs>=1.4.2
mkdocs-material>=8.5.11
pymdown-extensions>=9.9.2
markdown>=3.4.1
jinja2>=3.1.2
pyyaml>=6.0
python-dateutil>=2.8.2

# Documentation building
pandoc>=2.19.2
weasyprint>=58.1
mkdocs-pdf-export-plugin>=0.7.0
mkdocs-awesome-pages-plugin>=2.8.0
mkdocs-git-revision-date-localized-plugin>=1.2.0
mkdocs-minify-plugin>=0.6.2
mkdocs-redirects>=1.2.0
mkdocs-section-index>=0.3.5
mkdocs-simple-hooks>=0.1.3
mkdocs-static-i18n>=0.53
mkdocs-with-pdf>=0.9.0

# Documentation validation
markdown-link-validator>=0.1.0
markdown-lint>=0.0.1
markdown-meta>=0.1.0
markdown-toc>=0.1.0
markdown-validator>=0.1.0

# Search and indexing
nltk>=3.8.1
whoosh>=2.7.4
elasticsearch>=8.5.3

# Translation
deep-translator>=1.11.4
googletrans>=4.0.0-rc1

# Development tools
black>=22.12.0
flake8>=6.0.0
isort>=5.11.4
mypy>=0.991
pytest>=7.2.1
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.1.0

# Testing
pytest>=7.2.1
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.1.0

# Documentation
sphinx>=5.3.0
sphinx-rtd-theme>=1.1.1
sphinx-autodoc-typehints>=1.19.5
sphinx-copybutton>=0.5.1
sphinx-design>=0.3.0
sphinx-external-toc>=0.3.1
sphinx-jinja>=1.1.1
sphinx-multiversion>=0.2.4
sphinx-panels>=0.6.0
sphinx-prompt>=1.5.0
sphinx-tabs>=3.4.1
sphinx-theme-builder>=0.2.0b1
sphinx-toolbox>=3.1.0
sphinxcontrib-applehelp>=1.0.2
sphinxcontrib-devhelp>=1.0.2
sphinxcontrib-htmlhelp>=2.0.0
sphinxcontrib-jsmath>=1.0.1
sphinxcontrib-qthelp>=1.0.3
sphinxcontrib-serializinghtml>=1.1.5
sphinxcontrib-spelling>=7.7.0
sphinxcontrib-websupport>=1.2.4

# Utilities
click>=8.1.3
colorama>=0.4.6
humanize>=4.6.0
progress>=1.6
python-slugify>=8.0.1
rich>=13.6.0
typer>=0.7.0

# Network analysis
pcapy-ng>=1.0.9

# Core dependencies
arpguard==1.0.0
pyyaml>=5.4.1,<6.0.0
netifaces>=0.10.9,<0.11.0
psutil>=5.8.0,<6.0.0
requests>=2.27.0,<2.28.0

# UI components
PyQt5==5.15.2
pyqtgraph==0.12.4
pyqtwebengine==5.15.2

# Network analysis
scapy>=2.4.5,<2.5.0
numpy>=1.19.5,<1.20.0
netaddr>=0.8.0,<0.9.0

# Testing and development
pytest>=6.2.5,<7.0.0
pytest-cov>=2.12.1,<3.0.0
mock>=4.0.3,<5.0.0

# Core dependencies
pyyaml>=5.4.1,<6.0.0
netifaces>=0.10.9,<0.11.0
psutil>=5.8.0,<6.0.0

# UI components
PyQt5==5.15.2
pyqtgraph>=0.12.1
pyqtwebengine>=5.15.2

# Network analysis
scapy>=2.4.5
numpy>=1.19.5

# Testing and development
pytest>=6.2.5
pytest-cov>=2.12.1
mock>=4.0.3 